refactor(services): extract CodingService from Orchestrator
Move assign_codes and its four private helpers — _check_coding_deadline,
_traverse_coding_level, _pick_code_indices and _judge_code_level — onto a
new CodingService in qfa.services.coding, together with the module-level
coding helpers _ScoredCode, _as_whole_percentage,
_combine_rejected_explanations and the NO_CODING_* constants. The moved
code is byte-identical; Orchestrator loses 368 lines and gains nothing.

Per ADR-017 the service has no base class. It names the three
dependencies it actually uses — the primary LLM connection, the
anonymiser and the shared LLMCallExecutor — and takes no judge client at
all, which makes #258's exclusion of the per-level coding judge
structural rather than a convention.

qfa.api.composition grows build_services, returning both services as a
ServiceGraph so they share one executor and one anonymiser;
build_orchestrator stays as the narrow entry point for scripts and
notebooks and now delegates to it. The lifespan publishes
app.state.coding_service, get_coding_service reads it back, and the
assign-codes handler annotates against CodingService instead of
Orchestrator.

The coding tests move to tests/services/test_coding.py unchanged,
constructing the real executor over the existing FakeLLMPort /
FakeAnonymizer doubles; the exact rendered message pinned by #256 still
asserts the same string. New tests cover the executor sharing and the
app.state publication. The assign-codes API and e2e assertions are
untouched — only the import path of NO_CODING_EMPTY_CONTENT_EXPLANATION
and the test app's wiring changed.
