+++ src/qfa/api/app.py
@@ -612,11 +612,13 @@ def _make_lifespan(llm_factory: LLMFactory):
            ``build_services``) so its construction is visible in
            startup logs before any traffic arrives.
         5. Delegate to :func:`qfa.api.composition.build_services`
-           to assemble the application services — it also registers
-           custom LiteLLM model prices needed for ``completion_cost()``.
+           to assemble the application services over one shared
+           executor — it also registers custom LiteLLM model prices
+           needed for ``completion_cost()``.
         6. Publish ``orchestrator``, ``sensitivity_service``,
-           ``api_keys``, ``settings``, and ``usage_repo`` on
-           ``app.state`` for routes/middleware to read.
+           ``coding_service``, ``api_keys``, ``settings``, and
+           ``usage_repo`` on ``app.state`` for routes/middleware to
+           read.
 
         On shutdown the only resource that needs explicit cleanup is the
         DB engine's connection pool; everything else is plain Python
@@ -691,6 +693,7 @@ def _make_lifespan(llm_factory: LLMFactory):
         # One provider per use-case service (ADR-017): each route reads the
         # single service it needs off app.state.
         app.state.sensitivity_service = services.sensitivity
+        app.state.coding_service = services.coding
         app.state.settings = settings
         app.state.usage_repo = usage_repo
